Move relevant types to internal.d.ts
