+++ packages/gasket-plugin-docs/lib/utils/config-set-builder.js
@@ -89,7 +89,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Look up all doc files for a module
-   * @type {import('../index')._findAllFiles}
+   * @type {import('../internal')._findAllFiles}
    */
   async _findAllFiles(moduleData, docsSetup, link, sourceRoot) {
     if (!sourceRoot) return [];
@@ -135,7 +135,7 @@ class DocsConfigSetBuilder {
    * Divides global and local transforms from a docsSetup. Global transforms are
    * added to the top-level set. Local transforms will be added to the module's
    * docConfig.
-   * @type {import('../index')._segregateTransforms}
+   * @type {import('../internal')._segregateTransforms}
    */
   _segregateTransforms(transforms) {
     return transforms.reduce((acc, tx) => {
@@ -150,7 +150,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Constructs the DocsConfig for a module based on its info and docsSetup
-   * @type {import('../index')._buildDocsConfig}
+   * @type {import('../internal')._buildDocsConfig}
    */
   async _buildDocsConfig(moduleData, docsSetup = {}, overrides = {}) {
     const {
@@ -193,7 +193,7 @@ class DocsConfigSetBuilder {
   /**
    * Flattens all detail types from plugins' metadata. Add a from property with
    * name of parent plugin.
-   * @type {import('../index')._flattenDetails}
+   * @type {import('../internal')._flattenDetails}
    */
   _flattenDetails(type) {
     const arr = [];
@@ -220,7 +220,7 @@ class DocsConfigSetBuilder {
    * Adds additional docsSetup for modules, merging duplicates with a first in
    * wins approach. When a module is then add to be configured, a docSetup will
    * be looked up from what's been added by plugins here.
-   * @type {import('../index')._addModuleDocsSetup}
+   * @type {import('../internal')._addModuleDocsSetup}
    */
   _addModuleDocsSetup(moduleDocsSetup) {
     defaultsDeep(this._moduleDocsSetups, moduleDocsSetup);
@@ -228,7 +228,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for the App
-   * @type {import('../index').addApp}
+   * @type {import('../internal').addApp}
    */
   async addApp(moduleData, docsSetup) {
     // If docsSetup is passed, stick with it. Or, see if gasket.docsSetup in
@@ -247,7 +247,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for a plugin
-   * @type {import('../index').addPlugin}
+   * @type {import('../internal').addPlugin}
    */
   async addPlugin(pluginData, docsSetup) {
     if (this._plugins.find((p) => p.metadata === pluginData)) return;
@@ -282,7 +282,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for multiple plugins if not already added
-   * @type {import('../index').addPlugins}
+   * @type {import('../internal').addPlugins}
    */
   async addPlugins(pluginDatas) {
     await Promise.all(pluginDatas.map((p) => this.addPlugin(p)));
@@ -290,7 +290,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for a preset
-   * @type {import('../index').addPreset}
+   * @type {import('../internal').addPreset}
    */
   async addPreset(presetData, docsSetup) {
     if (this._presets.find((p) => p.metadata === presetData)) return;
@@ -315,7 +315,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for multiple presets if not already added
-   * @type {import('../index').addPresets}
+   * @type {import('../internal').addPresets}
    */
   async addPresets(presetDatas) {
     await Promise.all(presetDatas.map((p) => this.addPreset(p)));
@@ -323,7 +323,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for a module
-   * @type {import('../index').addModule}
+   * @type {import('../internal').addModule}
    */
   async addModule(moduleData, docsSetup) {
     if (this._modules.find((p) => p.metadata === moduleData)) return;
@@ -348,7 +348,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Add DocsConfig to the set for multiple modules if not already added
-   * @type {import('../index').addModules}
+   * @type {import('../internal').addModules}
    */
   async addModules(moduleDatas) {
     await Promise.all(moduleDatas.map((p) => this.addModule(p)));
@@ -356,7 +356,7 @@ class DocsConfigSetBuilder {
 
   /**
    * Picks out properties to return as the config set
-   * @type {import('../index').getConfigSet}
+   * @type {import('../internal').getConfigSet}
    */
   getConfigSet() {
     const detailDocsConfigs = detailDocsTypes.reduce(
